# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List my projects
      responses:
        "200":
          description: My projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project with entries
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: patchProject
      tags: [projects]
      summary: Update project title or visibility
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/entries:
    post:
      operationId: createEntry
      tags: [projects]
      summary: Add an entry to a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /projects/{projectId}/entries/{entryId}:
    delete:
      operationId: deleteEntry
      tags: [projects]
      summary: Delete an entry
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/export:
    get:
      operationId: exportProject
      tags: [projects]
      summary: Export project as JSON
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project export
          content:
            application/json:
              schema:
                type: object
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProjectSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        provider:
          type: string
        summary:
          type: string
          nullable: true
        isPublic:
          type: boolean
        createdAt:
          type: string
        entryCount:
          type: integer
      required: [id, title, category, provider, isPublic, createdAt, entryCount]
    NewProjectInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [rag-dataset, fine-tune, document, other]
        provider:
          type: string
      required: [title, category, provider]
    ProjectPatch:
      type: object
      properties:
        title:
          type: string
        isPublic:
          type: boolean
    NewEntryInput:
      type: object
      properties:
        inputUrl:
          type: string
      required: [inputUrl]
...
    EntryDescription:
      type: object
      properties:
        label:
          type: string
        text:
          type: string
      required: [label, text]
    Entry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        inputUrl:
          type: string
        aiQuestion:
          type: string
          nullable: true
        descriptions:
          type: array
          items:
            $ref: "#/components/schemas/EntryDescription"
        createdAt:
          type: string
      required: [id, projectId, inputUrl, descriptions, createdAt]
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        provider:
          type: string
        summary:
          type: string
          nullable: true
        isPublic:
          type: boolean
        createdAt:
          type: string
        role:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
      required: [id, title, category, provider, isPublic, createdAt, role, entries]
...