# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Entry } from './entry';

export interface ProjectDetail {
  id: number;
  title: string;
  category: string;
  provider: string;
  summary?: string | null;
  isPublic: boolean;
  createdAt: string;
  role: string;
  entries: Entry[];
}